Chart ticks.

+++ sources/Waves.UI.Charts/Drawing/Primitives/Interfaces/IWavesDrawingObject.cs
@@ -31,7 +31,7 @@ public interface IWavesDrawingObject
     /// <summary>
     ///     Gets or sets stroke thickness.
     /// </summary>
-    float StrokeThickness { get; set; }
+    double StrokeThickness { get; set; }
 
     /// <summary>
     ///     Gets or sets fill.
